+++ Project.toml
@@ -15,6 +15,7 @@ Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 Krylov = "ba0b0d4f-ebba-5204-a429-3ac8c609bfb7"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+LinearOperators = "5c8ed15e-5a4c-59e4-a42b-c7e8811fb125"
 MatrixMarket = "4d4711f2-db25-561a-b6b3-d35e7d4047d3"
 Metis = "2679e427-3c69-5b7f-982b-ece356f1e94b"
 NLsolve = "2774e3e8-f4cf-5e23-947b-6d7e65073b56"
@@ -35,6 +36,7 @@ Graphs = "1.12.0"
 Ipopt = "1.7.2"
 JuMP = "1.24.0"
 Krylov = "0.9.9"
+LinearOperators = "2.9.0"
 Metis = "1.5.0"
 NLsolve = "4.5.1"
 Optim = "1.9.4"
