+++ src/homological_programming/HomologicalProgramming.jl
@@ -2,14 +2,18 @@ module HomologicalProgramming
 
 using Reexport
 
+include("MPC.jl")
 include("CellularSheaves.jl")
-include("SheafNodes.jl")
-include("DistributedSheaves.jl")
-include("ThreadedSheaves.jl")
+#include("SheafNodes.jl")
+#include("DistributedSheaves.jl")
+#include("ThreadedSheaves.jl")
+include("HomologicalPrograms.jl")
 
+@reexport using .MPC
 @reexport using .CellularSheaves
-@reexport using .SheafNodes
-@reexport using .DistributedSheaves
-@reexport using .ThreadedSheaves
+#@reexport using .SheafNodes
+#@reexport using .DistributedSheaves
+#@reexport using .ThreadedSheaves
+@reexport using .HomologicalPrograms
 
 end
